requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorclass==2.2.2
Django==4.1.4
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
packaging==22.0
pip-upgrader==1.4.15
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
terminaltables==3.1.10
urllib3==1.26.13
